update to reflect revisions
