+++ src/bulk_fluxes/generate_error_stats_table.py
@@ -1,3 +1,6 @@
+"""
+Generates the data displayed in Table 2
+"""
 import matplotlib.pyplot as plt
 from scipy.ndimage import median_filter
 import numpy as np
