+++ src/bulk_variables/plot_internal_temp.py
@@ -1,3 +1,6 @@
+"""
+Generates Figure 3a
+"""
 import pandas as pd
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
